requirements.txt
django
google-genai
cloudinary
django-cloudinary-storage
psycopg2-binary
python-dotenv
pillow
gunicorn

